[FE] modify: 여행지 목록 API 위치 변경
